bug fux

+++ judge/views/problem.py
@@ -261,9 +261,9 @@ class ProblemDetail(ProblemMixin, SolvedProblemMixin, CommentedDetailView):
         def median(l,r,data): #provides index and value of the median of some range of the data
             size = r-l+1
             if size % 2 == 1:
-                return l+size/2,data[l+size/2]
+                return l+size/2,data[l+int(size/2)]
             else:
-                return l+size/2+0.5,(data[l+size/2]+data[l+size/2+1])/2
+                return l+size/2+0.5,(data[l+int(size/2)]+data[l+int(size/2)+1])/2
 
         #box and whisker plot data
         q2 = median(0,len(all_votes)-1,all_votes) #median
